move logger to core
