+++ log/logger.go
@@ -9,6 +9,8 @@ import (
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/pkgerrors"
+
+	corelog "cosmossdk.io/core/log"
 )
 
 func init() {
@@ -32,34 +34,7 @@ const ModuleKey = "module"
 // ContextKey is used to store the logger in the context.
 var ContextKey struct{}
 
-// Logger is the Cosmos SDK logger interface.
-// It maintains as much backward compatibility with the CometBFT logger as possible.
-// All functionalities of the logger are available through the Impl() method.
-type Logger interface {
-	// Info takes a message and a set of key/value pairs and logs with level INFO.
-	// The key of the tuple must be a string.
-	Info(msg string, keyVals ...any)
-
-	// Warn takes a message and a set of key/value pairs and logs with level WARN.
-	// The key of the tuple must be a string.
-	Warn(msg string, keyVals ...any)
-
-	// Error takes a message and a set of key/value pairs and logs with level ERR.
-	// The key of the tuple must be a string.
-	Error(msg string, keyVals ...any)
-
-	// Debug takes a message and a set of key/value pairs and logs with level DEBUG.
-	// The key of the tuple must be a string.
-	Debug(msg string, keyVals ...any)
-
-	// With returns a new wrapped logger with additional context provided by a set.
-	With(keyVals ...any) Logger
-
-	// Impl returns the underlying logger implementation.
-	// It is used to access the full functionalities of the underlying logger.
-	// Advanced users can type cast the returned value to the actual logger.
-	Impl() any
-}
+type Logger = corelog.Logger
 
 // WithJSONMarshal configures zerolog global json encoding.
 func WithJSONMarshal(marshaler func(v any) ([]byte, error)) {
