+++ runtime/environment.go
@@ -2,8 +2,8 @@ package runtime
 
 import (
 	"cosmossdk.io/core/appmodule"
+	"cosmossdk.io/core/log"
 	"cosmossdk.io/core/store"
-	"cosmossdk.io/log"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 )
